Move CreateEventFragment.kt and CreateUserFragment.kt to fragments folder
