+++ app/src/main/res/layout/fragment_create_event.xml
@@ -4,7 +4,7 @@
     android:layout_height="match_parent"
     xmlns:tools="http://schemas.android.com/tools"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".CreateEventFragment">
+    tools:context=".fragments.createevent.CreateEventFragment">
 
     <TextView
         android:id="@+id/event_name_label"
