+++ app/src/main/res/layout/fragment_create_user.xml
@@ -4,7 +4,7 @@
     android:layout_height="match_parent"
     xmlns:tools="http://schemas.android.com/tools"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".CreateUserFragment">
+    tools:context=".fragments.createuser.CreateUserFragment">
 
     <ImageView
         android:id="@+id/user_iv"
